chore: More progress
